完成了登录和个人信息

+++ xzcp/src/cn/xzcp/bean/Score.java
@@ -6,37 +6,45 @@ public class Score {
 
 	private int scoreId;
 	private int scoreStudentid;
-	private int scoreScore;
+	private double scoreScore;
 	private Date scoreDate;
+
 	@Override
 	public String toString() {
 		return "Score [scoreId=" + scoreId + ", scoreStudentid=" + scoreStudentid + ", scoreScore=" + scoreScore
 				+ ", scoreDate=" + scoreDate + "]";
 	}
+
 	public int getScoreId() {
 		return scoreId;
 	}
+
 	public void setScoreId(int scoreId) {
 		this.scoreId = scoreId;
 	}
+
 	public int getScoreStudentid() {
 		return scoreStudentid;
 	}
+
 	public void setScoreStudentid(int scoreStudentid) {
 		this.scoreStudentid = scoreStudentid;
 	}
-	public int getScoreScore() {
+
+	public double getScoreScore() {
 		return scoreScore;
 	}
-	public void setScoreScore(int scoreScore) {
+
+	public void setScoreScore(double scoreScore) {
 		this.scoreScore = scoreScore;
 	}
+
 	public Date getScoreDate() {
 		return scoreDate;
 	}
+
 	public void setScoreDate(Date scoreDate) {
 		this.scoreDate = scoreDate;
 	}
-	
 
 }
